% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_vec}
\alias{dir_vec}
\title{L2 Direction of a Vector}
\usage{
dir_vec(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with L2 norm equal to 1.
}
\description{
Computes the normalized (L2 direction) vector.
}
\keyword{internal}
