% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncation_interval.R
\name{minus_quad_ineq}
\alias{minus_quad_ineq}
\title{Implement the minus operation for two quadratic inequalities}
\usage{
minus_quad_ineq(quad1, quad2)
}
\arguments{
\item{quad1, }{first inequality}

\item{quad2, }{second inequality}
}
\description{
Implement the minus operation for two quadratic inequalities
}
\keyword{internal}
