% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncation_interval.R
\name{norm_phi_canonical_panel}
\alias{norm_phi_canonical_panel}
\title{Quadratic Form of ||z_it(ϕ) - ClusterMean_k(ϕ)||² for Canonical Form}
\usage{
norm_phi_canonical_panel(rq, i, k, last_centroids, weighted_deltahat)
}
\arguments{
\item{rq}{List from \code{required_quantities()} (matrix-based)}

\item{i}{Integer; unit index}

\item{k}{Integer; cluster index}

\item{last_centroids}{K x P matrix of centroids from previous k-means step}

\item{weighted_deltahat}{K-length vector of weighted δₖ}
}
\value{
A list with \code{quad}, \code{linear}, \code{constant} terms
}
\description{
Represent
\deqn{||[z(ϕ)]_{it} - T^{-1} ∑_{t} ∑_{j} w_j^{(m-1)}(k) [z(ϕ)]_{jt}||²}
as a quadratic function in ϕ.
}
\keyword{internal}
