% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_vec}
\alias{norm_vec}
\title{L2 Norm of a Vector}
\usage{
norm_vec(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar representing the L2 norm of \code{x}.
}
\description{
Computes the Euclidean (L2) norm of a numeric vector.
}
\keyword{internal}
