% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{panel_kmeans_estimation}
\alias{panel_kmeans_estimation}
\title{Panel K-means Estimation with Optional BIC Selection (Parallelized)}
\usage{
panel_kmeans_estimation(
  Z,
  id,
  time,
  K = NULL,
  Kmax = NULL,
  Ninit = 10,
  iter.max = 10,
  n_cores = 1
)
}
\arguments{
\item{Z}{NT x P matrix of variables for clustering (typically stacked by time)}

\item{id}{NT-length vector of unit identifiers (repeating across time)}

\item{time}{NT-length vector of time identifiers (repeating across units)}

\item{K}{Integer; number of clusters. Required unless Kmax is provided.}

\item{Kmax}{Optional; if provided, perform BIC-based selection from 2 to Kmax clusters.}

\item{Ninit}{Number of random initializations}

\item{iter.max}{Maximum number of Lloyd iterations}

\item{n_cores}{Number of cores for parallelization (default: 1, i.e., no parallel)}
}
\value{
A list with: clusters, centers, objectives, iter, final_cluster, BIC_selected_K (if Kmax is used)
}
\description{
This function performs K-means clustering on NT x P matrix panel data. If `Kmax` is provided,
it performs model selection using BIC over 2 to `Kmax` clusters. Otherwise, clustering is
performed using the provided `K`. Parallelization is used for random initializations.
}
