% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncation_interval.R
\name{required_quantities}
\alias{required_quantities}
\title{Prepare Quantities for Truncation Set in Two-Group Inference}
\usage{
required_quantities(Z, id, time, k1, k2, gamma, OmegaHat)
}
\arguments{
\item{Z}{NT x P matrix of stacked panel data (P can be 1)}

\item{id}{Vector of unit identifiers (length NT, integers)}

\item{time}{Vector of time identifiers (length NT, integers)}

\item{k1}{First cluster index (scalar)}

\item{k2}{Second cluster index (scalar)}

\item{gamma}{Vector of length N with estimated cluster labels}

\item{OmegaHat}{KP x KP long-run variance-covariance matrix}
}
\value{
A list of required quantities for computing truncation intervals
}
\description{
This internal function computes key quantities used for selective inference in 
comparing cluster means. Inputs are assumed to be already in matrix form with
numeric `id` and `time` vectors of length NT.
}
\keyword{internal}
