% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_cl}
\alias{same_cl}
\title{Check for Cluster Label Equivalence up to Permutation}
\usage{
same_cl(cl1, cl2, K)
}
\arguments{
\item{cl1}{A vector of cluster labels.}

\item{cl2}{Another vector of cluster labels.}

\item{K}{Number of clusters.}
}
\value{
Logical; \code{TRUE} if clusterings are identical up to permutation.
}
\description{
Compares two clustering vectors to determine if they differ only by a permutation.
}
\keyword{internal}
